refactor/#169: 핸들러 분리 및 훅 분리
